NotificationManager
Add NotificationManager
