+++ MyMusicApp/Controllers/OnboardingVC/FirstScreenOnboardingVC.swift
@@ -102,10 +102,6 @@ class FirstScreenOnboardingVC: UIViewController {
         view.backgroundColor = .black
         
         setupUI()
-        userNotificationPermission()
-        sendNotification()
-        
-        notificationCenter.delegate = self
     }
     
     // MARK: - Methods
@@ -114,41 +110,7 @@ class FirstScreenOnboardingVC: UIViewController {
     }
 }
 
-// MARK: - Notification
-public let notificationCenter = UNUserNotificationCenter.current()
-
-public func userNotificationPermission() {
-    //ask the permission from user
-    notificationCenter.requestAuthorization(options: [.alert, .sound, .badge]) { (granted, error) in
-        guard granted else { return }
-        notificationCenter.getNotificationSettings { (settings) in
-            print(settings)
-            guard settings.authorizationStatus == .authorized else { return }
-        }
-    }
-}
-
-public func sendNotification() {
-    //create notification content
-    let content = UNMutableNotificationContent()
-    content.title = "Greetings! 🥰"
-    content.body = " You can turn on/off notifications in app settings"
-    content.sound = UNNotificationSound.default
-    
-    //create the notification trigger
-    let trigger = UNTimeIntervalNotificationTrigger(timeInterval: 5, repeats: false)
-    
-    //create the request
-    let request = UNNotificationRequest(identifier: "notification", content: content, trigger: trigger)
-    
-    //register the notification request
-    notificationCenter.add(request) { (error) in
-        print(error?.localizedDescription as Any)
-    }
-}
-
-
-extension FirstScreenOnboardingVC: UNUserNotificationCenterDelegate {
+extension FirstScreenOnboardingVC {
     
     // MARK: - UI Setup
     func setupUI() {
@@ -223,19 +185,5 @@ extension FirstScreenOnboardingVC: UNUserNotificationCenterDelegate {
             getStartedButton.widthAnchor.constraint(equalToConstant: 280),
             getStartedButton.heightAnchor.constraint(equalToConstant: 50),
         ])
-    }
-    
-    // MARK: - Notifications delegate
-    func userNotificationCenter(_ center: UNUserNotificationCenter, willPresent notification: UNNotification, withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
-        completionHandler([.alert, .sound, .badge])
-        print(#function)
-    }
-    
-    func userNotificationCenter(_ center: UNUserNotificationCenter, didReceive response: UNNotificationResponse, withCompletionHandler completionHandler: @escaping () -> Void) {
-        print(#function)
-        //user tap on badge and go to the screen we need
-        let vc = AccountViewController()
-        vc.modalPresentationStyle = .fullScreen
-        present(vc, animated: true)
     }
 }
